+++ test/MGNZ.Squidex.CLI.Tests/Routing/RouteMetadataBuilderUnitTests.cs
@@ -1,11 +1,11 @@
-namespace MGNZ.Squidex.CLI.Tests.CLI
+namespace MGNZ.Squidex.CLI.Tests.Routing
 {
   using System.Collections.Generic;
   using System.Reflection;
 
   using FluentAssertions;
 
-  using MGNZ.Squidex.CLI.Common.CLI;
+  using MGNZ.Squidex.CLI.Common.Routing;
   using MGNZ.Squidex.CLI.Tests.Platform;
 
   using Xunit;
