+++ test/MGNZ.Squidex.CLI.Tests/Routing/RouteOptionsParserUnitTests.cs
@@ -1,11 +1,11 @@
-namespace MGNZ.Squidex.CLI.Tests.CLI
+namespace MGNZ.Squidex.CLI.Tests.Routing
 {
   using System.Collections.Generic;
   using System.Linq;
 
   using FluentAssertions;
 
-  using MGNZ.Squidex.CLI.Common.CLI;
+  using MGNZ.Squidex.CLI.Common.Routing;
   using MGNZ.Squidex.CLI.Common.Commands;
   using MGNZ.Squidex.CLI.Tests.Platform;
